Always use '$' as the mutation id separator in concatenated id strings
Rather than using '$' in URL params and ',' in local storage as we were
doing before this commit.

Note that this commit will invalidate the user's local storage of the
ids of active mutations, meaning that their latest selection of
mutations will be wiped clean the next time they load up the app with
this commit in it and without a URL parameter specifying alternate
mutation ids. This is due to the app no longer recognizing ',' as an id
separator and so attempting to match the entire string of local storage
ids to a single mutation id, which won't be found and so the selection
is wiped clean.